% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{carpentries_colors}
\alias{carpentries_colors}
\title{A Vector containing the color palette for The Carpentries}
\format{
An object of class \code{character} of length 14.
}
\usage{
carpentries_colors
}
\description{
A Vector containing the color palette for The Carpentries
}
\keyword{datasets}
