% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{carpentries_pal}
\alias{carpentries_pal}
\title{Carpentries color palette (discrete)}
\usage{
carpentries_pal()
}
\value{
a function that generates a palette of up to 14 colors.
}
\description{
This palette includes 14 colors.
}
\examples{

my_palette <- carpentries_pal()
my_palette(3)

}
