% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_carpentries}
\alias{theme_carpentries}
\title{Carpentries GGPlot2 Theme}
\usage{
theme_carpentries()
}
\description{
Carpentries GGPlot2 Theme
}
\examples{

mtcars2 <- within(mtcars, {
  vs <- factor(vs, labels = c("V-shaped", "Straight"))
  am <- factor(am, labels = c("Automatic", "Manual"))
  cyl  <- factor(cyl)
  gear <- factor(gear)
})

p1 <- ggplot(mtcars2) +
  geom_point(aes(x = wt, y = mpg, colour = gear)) +
  labs(
    title = "Fuel economy declines as weight increases",
    subtitle = "(1973-74)",
    caption = "Data from the 1974 Motor Trend US magazine.",
    tag = "Figure 1",
    x = "Weight (1000 lbs)",
    y = "Fuel economy (mpg)",
    colour = "Gears"
  )

}
